Split up the ifdformat module
Reimplement tag_value_new() as TagValue::new(), and put ToCsv in its own
module.
